Move JS sources according to 6.1 changes
