+++ src/Resources/app/storefront/src/payolution-installment/payone-payment.payolution-installment.js
@@ -1,8 +1,8 @@
 /* eslint-disable import/no-unresolved */
 
-import Plugin from 'src/script/plugin-system/plugin.class';
-import HttpClient from 'src/script/service/http-client.service';
-import PageLoadingIndicatorUtil from 'src/script/utility/loading-indicator/page-loading-indicator.util';
+import Plugin from 'src/plugin-system/plugin.class';
+import HttpClient from 'src/service/http-client.service';
+import PageLoadingIndicatorUtil from 'src/utility/loading-indicator/page-loading-indicator.util';
 
 export default class PayonePaymentPayolutionInstallment extends Plugin {
     init() {
