+++ src/Resources/app/storefront/src/script/payone-payment.js
