+++ src/Resources/app/storefront/src/style/payone-payment.scss
